3422: Solver: Errorhandling Mid, Alphasum with umlauts Android FilePicker: sort files and directories Skin: correct selection of skins and avoid usage of file External request: correct start of CB some internal restructuring

+++ CB_UI_Base/src/CB_UI_Base/Global.java
@@ -12,7 +12,7 @@ public abstract class Global {
     public static boolean forcePhone = true;
     public static boolean useSmallSkin = false;
     public static DisplayType displayType = DisplayType.Normal;
-    public static double displayDensity = 1;
+    public static float displayDensity = 1;
     protected static Global Instance;
     private static Clipboard defaultClipBoard;
